Modified structure of webtool python codes
